+++ table/partitioned_filter_block.cc
@@ -82,21 +82,26 @@ PartitionedFilterBlockReader::PartitionedFilterBlockReader(
     : FilterBlockReader(contents.data.size(), stats, _whole_key_filtering),
       prefix_extractor_(prefix_extractor),
       comparator_(comparator),
-      table_(table),
-      block_cache_(table_->rep_->table_options.block_cache.get()){
+      table_(table) {
   idx_on_fltr_blk_.reset(new Block(std::move(contents),
                                    kDisableGlobalSequenceNumber,
                                    0 /* read_amp_bytes_per_bit */, stats));
 }
 
 PartitionedFilterBlockReader::~PartitionedFilterBlockReader() {
-  // The destructor migh be called via cache evict after the table is deleted.
-  // We should avoid using table_ pointer in destructor then.
-  if (block_cache_) {
+  {
     ReadLock rl(&mu_);
     for (auto it = handle_list_.begin(); it != handle_list_.end(); ++it) {
-      block_cache_->Release(*it);
+      table_->rep_->table_options.block_cache.get()->Release(*it);
     }
+  }
+  char cache_key[BlockBasedTable::kMaxCacheKeyPrefixSize + kMaxVarint64Length];
+  for (auto it = filter_block_set_.begin(); it != filter_block_set_.end();
+       ++it) {
+    auto key = BlockBasedTable::GetCacheKey(table_->rep_->cache_key_prefix,
+                                            table_->rep_->cache_key_prefix_size,
+                                            *it, cache_key);
+    table_->rep_->table_options.block_cache.get()->Erase(key);
   }
 }
 
@@ -111,8 +116,6 @@ bool PartitionedFilterBlockReader::KeyMayMatch(
   if (UNLIKELY(idx_on_fltr_blk_->size() == 0)) {
     return true;
   }
-  // This is the user key vs. the full key in the partition index. We assume
-  // that user key <= full key
   auto filter_handle = GetFilterPartitionHandle(*const_ikey_ptr);
   if (UNLIKELY(filter_handle.size() == 0)) {  // key is out of range
     return false;
@@ -127,7 +130,7 @@ bool PartitionedFilterBlockReader::KeyMayMatch(
     return res;
   }
   if (LIKELY(filter_partition.IsSet())) {
-    filter_partition.Release(block_cache_);
+    filter_partition.Release(table_->rep_->table_options.block_cache.get());
   } else {
     delete filter_partition.value;
   }
@@ -159,7 +162,7 @@ bool PartitionedFilterBlockReader::PrefixMayMatch(
     return res;
   }
   if (LIKELY(filter_partition.IsSet())) {
-    filter_partition.Release(block_cache_);
+    filter_partition.Release(table_->rep_->table_options.block_cache.get());
   } else {
     delete filter_partition.value;
   }
@@ -187,7 +190,8 @@ PartitionedFilterBlockReader::GetFilterPartition(Slice* handle_value,
   auto s = fltr_blk_handle.DecodeFrom(handle_value);
   assert(s.ok());
   const bool is_a_filter_partition = true;
-  if (LIKELY(block_cache_ != nullptr)) {
+  auto block_cache = table_->rep_->table_options.block_cache.get();
+  if (LIKELY(block_cache != nullptr)) {
     bool pin_cached_filters =
         GetLevel() == 0 &&
         table_->rep_->table_options.pin_l0_filter_and_index_blocks_in_cache;
@@ -202,15 +206,18 @@ PartitionedFilterBlockReader::GetFilterPartition(Slice* handle_value,
     }
     auto filter =
         table_->GetFilter(fltr_blk_handle, is_a_filter_partition, no_io);
-    if (pin_cached_filters && filter.IsSet()) {
-      WriteLock wl(&mu_);
-      std::pair<uint64_t, FilterBlockReader*> pair(fltr_blk_handle.offset(),
-                                                   filter.value);
-      auto succ = filter_cache_.insert(pair).second;
-      if (succ) {
-        handle_list_.push_back(filter.cache_handle);
-      }  // Otherwise it is already inserted by a concurrent thread
-      *cached = true;
+    if (filter.IsSet()) {
+      filter_block_set_.insert(fltr_blk_handle);
+      if (pin_cached_filters) {
+        WriteLock wl(&mu_);
+        std::pair<uint64_t, FilterBlockReader*> pair(fltr_blk_handle.offset(),
+                                                     filter.value);
+        auto succ = filter_cache_.insert(pair).second;
+        if (succ) {
+          handle_list_.push_back(filter.cache_handle);
+        }  // Otherwise it is already inserted by a concurrent thread
+        *cached = true;
+      }
     }
     return filter;
   } else {